Added support for multiple Aspect Ratios
Initial resolution will now be dynamically set according
to the aspect ratio of the display. Resolution settings should
also change according to the aspect ratio so they cannot be set
to an undesired or unusable state.

ALSO: there is some extra stuff in here that I dont know how to exclude
from the commit, like the settings file and the iml file