Fixed bug that would not update the adenture state's scale and would skew the mouse position
Adjusted the target resulution

Set scale to round to the nearest integer and added a variable to define autoscalling.

Fixed bug in xbox controller where button press would trigger for both press and release

removed stamina and magic, introduced intelligence (magic) and endurance and will power (defences)

updated mouse lsiteners to only take the new x and why variables

Created the base for the item system.

Created a character class to be used as the holder for the player's info

created an inventory class to hod lsts of items for the character,