Added a new mini number font to use for health "gibs"
Added a sprite system.

Added a simple sword swing sprite to accompany slash

Upfated the tmx file to seperate background and forground

Added new sounds and added them for swing, hurt and death

Updated the delta display

Upated the order in which entities are rendered - added an additional render method to entities to display identifiers (name, health, etC)

Added a system to precalc the draw psotions of entities so I can loop through them multiple times for layered drawing.

Added a basic framework for sprites and their "parent" extendable glasses. Global and local (on an entity or not).